% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vax_tibble.R
\name{build_vax_tibble}
\alias{build_vax_tibble}
\title{produce a tibble with CDC vaccine data}
\usage{
build_vax_tibble(csvpath)
}
\arguments{
\item{csvpath}{character(1) path to CSV as distributed at 
`https://data.cdc.gov/Vaccinations/COVID-19-Vaccinations-in-the-United-States-Jurisdi/unsk-b7fc`}
}
\description{
produce a tibble with CDC vaccine data
}
\examples{
path_12_29 = system.file("cdc/COVID-19_Vaccinations_in_the_United_States_Jurisdiction.csv",
  package="teachCovidData")
build_vax_tibble(path_12_29)  # retrieved 12/29, data to 12/22
}
