% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_vax_2022.12.21}
\alias{cdc_vax_2022.12.21}
\title{snapshot of CDC data on COVID vaccination rates}
\format{
data.frame
}
\usage{
cdc_vax_2022.12.21
}
\description{
snapshot of CDC data on COVID vaccination rates
}
\keyword{datasets}
