% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datadict.R
\name{datadict_app}
\alias{datadict_app}
\title{produce shiny app from processed data dictionary}
\usage{
datadict_app(pd)
}
\arguments{
\item{pd}{output of \code{\link{process_datadict}}}
}
\description{
produce shiny app from processed data dictionary
}
\examples{
if (interactive()) {
 pa = system.file("cdc/VACCDataDictionary_v36_12082022.xlsx", package="teachCovidData")
 pd = process_datadict(pa)
 datadict_app(pd)
}
}
