% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MS_counties_2022.12.21}
\alias{MS_counties_2022.12.21}
\title{snapshot of CDC county-level data on COVID incidence in Mississippi}
\format{
data.frame
}
\usage{
MS_counties_2022.12.21
}
\description{
snapshot of CDC county-level data on COVID incidence in Mississippi
}
\keyword{datasets}
