% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usafacts_confirmed.R
\name{new_cases_state}
\alias{new_cases_state}
\title{Assemble new cases by state}
\usage{
new_cases_state(statecode)
}
\arguments{
\item{statecode}{character(1)}
}
\description{
Assemble new cases by state
}
\examples{
ncms = new_cases_state("MS")
plot(ncms)
}
