% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usafacts_confirmed.R
\name{plot.tcov_deltaconf_daily_state}
\alias{plot.tcov_deltaconf_daily_state}
\title{plot daily new cases for a state}
\usage{
\method{plot}{tcov_deltaconf_daily_state}(x, y, ...)
}
\arguments{
\item{x}{instance of tcov_deltaconf_daily_state}

\item{y}{not used}

\item{\dots}{not used}
}
\description{
plot daily new cases for a state
}
