% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{s2p_get_cache}
\alias{s2p_get_cache}
\title{get the BiocFileCache for sars2pack -- code by Sean Davis, extracted
and modified by Vince Carey for teachCovidData -- specifically,
no automated age check for updating; user must ask for update, otherwise gets cached data}
\usage{
s2p_get_cache(...)
}
\arguments{
\item{cache}{character(1), the path to the cache directory. See [BiocFileCache::BiocFileCache()].}
}
\value{
a `BiocFileCache` instance
}
\description{
Set up or get location of BiocFileCache. By default,
this uses the name BiocFileCache.
}
\keyword{internal}
