% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usafacts_confirmed.R
\name{usafacts_confirmed}
\alias{usafacts_confirmed}
\title{get confirmed COVID-19 cases from USAfacts.org}
\usage{
usafacts_confirmed(update = FALSE)
}
\arguments{
\item{update}{logical(1) if TRUE will download the data, otherwise use (or populate then use) cache}
}
\description{
get confirmed COVID-19 cases from USAfacts.org
}
\examples{
cc = usafacts_confirmed() # confirmed cases by county
library(dplyr)
MSdata = cc |> filter(State=="MS")
head(MSdata[,1:10])
dates = lubridate::as_date(names(MSdata[-c(1:4)]))
cumulative = apply(MSdata[,-c(1:4)],2,sum)
library(ggplot2)
mydf1 = data.frame(cumulative, dates)
ggplot(mydf1, aes(x=dates, y=cumulative)) + geom_point() + ggtitle("Cumulative Confirmed Covid cases in MS")
incid = diff(cumulative)
ndate = dates[-1]
mydf2 = data.frame(incid, ndate)
ggplot(mydf2, aes(x=ndate, y=incid)) + geom_point() + ggtitle("New daily cases in MS")
}
