% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datadict.R
\name{process_datadict}
\alias{process_datadict}
\title{produce tables from xlsx data dictionary}
\usage{
process_datadict(path, quiet = TRUE, simplify_names = TRUE)
}
\arguments{
\item{path}{character(1) path to xlsx file with data dictionary in several sheets}

\item{quiet}{logical(1) if TRUE (default) suppress messages about column names from read_xlsx}

\item{simplify_names}{logical(1) if TRUE (default) remove inessential elements of sheet names}
}
\value{
list of data.frames
}
\description{
produce tables from xlsx data dictionary
}
\note{
Original xlsx file retrieved from \url{https://data.cdc.gov/api/views/unsk-b7fc/files/bfb13090-4148-4f48-9d5a-e1d1c2fb93b3?download=true&filename=DataDictionary_v36_12082022.xlsx} in early 2023.
}
\examples{
pa = system.file("cdc/VACCDataDictionary_v36_12082022.xlsx", package="teachCovidData")
pd = process_datadict(pa)
names(pd)
head(pd[[1]])
}
